requirements.txt
# Core dependencies
pygit2>=1.14.0
gitpython>=3.1.40
email-validator>=2.1.0

# AI providers
openai>=1.12.0
google-generativeai>=0.3.2
anthropic>=0.18.0

# Data processing
pandas>=2.1.4
numpy>=1.26.3
python-dateutil>=2.8.2

# HTTP and web
requests>=2.31.0
urllib3>=2.1.0
beautifulsoup4>=4.12.3
lxml>=5.0.0
aiohttp>=3.9.0
httpx>=0.26.0

# Text processing
tiktoken>=0.5.2

# Utilities
python-dotenv>=1.0.0
click>=8.1.7
tqdm>=4.66.1
colorama>=0.4.6

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3